client

+++ server/src/modules/users/user.repository.ts
@@ -14,39 +14,50 @@ export class UserRepository {
     return User.create(data);
   }
 
-    async findById(id: string) {
-    return User.findById(id).exec();
-  }
+  async findById(id: string) {
+  return User.findById(id)
+    .select('-passwordHash')
+    .populate({
+      path: 'notes',
+      options: {
+        sort: {
+          createdAt: -1,
+        },
+      },
+    })
+    .lean()
+    .exec();
+}
 
-  async findAll(page: number) {
-    const limit = 10;
-    const skip = (page - 1) * limit;
+async findAll(page: number) {
+  const limit = 10;
+  const skip = (page - 1) * limit;
 
-    const [docs, total] = await Promise.all([
-      User.find()
-        .select('-passwordHash')
-        .sort({ createdAt: -1 })
-        .skip(skip)
-        .limit(limit)
-        .lean()
-        .exec(),
+  const [docs, total] = await Promise.all([
+    User.find()
+      .select('-passwordHash')
+      .sort({ createdAt: -1 })
+      .skip(skip)
+      .limit(limit)
+      .lean()
+      .exec(),
 
-      User.countDocuments(),
-    ]);
+    User.countDocuments(),
+  ]);
 
-    return {
-      docs,
-      totalDocs: total,
-      page,
-      limit,
-      totalPages: Math.ceil(
-        total / limit,
-      ),
-      hasNextPage:
-        page < Math.ceil(total / limit),
-      hasPrevPage: page > 1,
-    };
-  }
+  return {
+    docs,
+    totalDocs: total,
+    page,
+    limit,
+    totalPages: Math.ceil(
+      total / limit,
+    ),
+    hasNextPage:
+      page < Math.ceil(total / limit),
+    hasPrevPage: page > 1,
+  };
+}
 
   async updateById(
     id: string,
@@ -54,6 +65,7 @@ export class UserRepository {
       userName?: string;
       email?: string;
       role?: 'user' | 'admin';
+      interests?: string[];
     },
   ) {
     return User.findByIdAndUpdate(
@@ -72,6 +84,11 @@ export class UserRepository {
   }
 
   async deleteById(id: string) {
+    // check core admin cannot be deleted
+    const user = await User.findById(id).exec();
+    if (user && user.role === 'admin' && user.email === process.env.ADMIN_EMAIL){
+      throw new Error('Cannot delete the core admin user');
+    }
     return User.findByIdAndDelete(id).exec();
   }
 }
